chore(shared-react): updated plugins to use shared-react/utils
