+++ plugins/jfrog-artifactory/package.json
@@ -30,6 +30,7 @@
     "@material-ui/core": "^4.9.13",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.45",
+    "@janus-idp/shared-react": "^1.2.0",
     "react-use": "^17.2.4"
   },
   "peerDependencies": {
