--- plugins/jfrog-artifactory/src/components/utils.tsx
@@ -1,24 +1,0 @@
-export function formatDate(date: string | number | Date) {
-  if (date === -1) {
-    return 'N/A';
-  }
-
-  const adjustedDate = typeof date === 'number' ? date * 1000 : date;
-  return new Date(adjustedDate).toLocaleString('en-US', {
-    timeZone: new Intl.DateTimeFormat().resolvedOptions().timeZone,
-    timeStyle: 'short',
-    dateStyle: 'medium',
-  });
-}
-
-export function formatSize(sizeInBytes: number) {
-  if (!sizeInBytes) {
-    // null or undefined
-    return 'N/A';
-  }
-
-  const i = Math.floor(Math.log(sizeInBytes) / Math.log(1024));
-  return `${Number((sizeInBytes / Math.pow(1024, i)).toFixed(2)) * 1} ${
-    ['B', 'kB', 'MB', 'GB', 'TB'][i]
-  }`;
-}
