+++ plugins/quay/package.json
@@ -31,6 +31,7 @@
     "@material-ui/core": "^4.12.2",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "4.0.0-alpha.61",
+    "@janus-idp/shared-react": "^1.2.0",
     "react-use": "^17.2.4"
   },
   "peerDependencies": {
