+++ plugins/shared-react/src/utils/date/format-date.test.ts
@@ -5,9 +5,29 @@ describe('formatDate', () => {
     expect(formatDate(undefined)).toEqual('N/A');
   });
 
-  it('should format date', () => {
-    expect(formatDate('2020-01-01T00:00:00.000Z')).toEqual(
-      'Jan 1, 2020, 12:00 AM',
-    );
+  it('formatDate should return formatted date', () => {
+    const formatDateVal = formatDate('2023-06-30T05:03:17.8253401Z');
+    expect(formatDateVal).toEqual('Jun 30, 2023, 5:03 AM');
+  });
+
+  it('formatDate should return formatted date for date', () => {
+    const formatDateVal = formatDate(new Date('2023-06-30T05:03:17.8253401Z'));
+    expect(formatDateVal).toEqual('Jun 30, 2023, 5:03 AM');
+  });
+
+  it('formatDate should return N/A for invalid date', () => {
+    const formatDateVal = formatDate(-1);
+    expect(formatDateVal).toEqual('N/A');
+  });
+
+  it('should correctly format a date when its a number', () => {
+    const date = formatDate(1678890069).normalize('NFKC');
+
+    expect(date).toBe('Mar 15, 2023, 2:21 PM');
+  });
+
+  it('formatDate should return N/A for empty string', () => {
+    const formatDateVal = formatDate('');
+    expect(formatDateVal).toEqual('N/A');
   });
 });
